A few new code experiments and some pcb files
Just messing around trying to figure out external interupts on the
attiny(for putting it to sleep), and a huge step(for me anyway) I have
the first board laid out. Kicad files and gerber files. Its not the
best(it may be a wreck), but as this project moves forward, hopefully
people actually are interested in adding to this, that would be awesome!